refactor(backend): update settings service position
