+++ backend/Zeus.Api.Infrastructure/Services/Settings/Integrations/DiscordSettingsProvider.cs
@@ -1,9 +1,7 @@
-using Microsoft.Extensions.Options;
-
-using Zeus.Api.Application.Interfaces.Services.Integrations;
+using Zeus.Api.Application.Interfaces.Services.Settings.Integrations;
 using Zeus.Api.Infrastructure.Settings.Integrations;
 
-namespace Zeus.Api.Infrastructure.Services.Integrations;
+namespace Zeus.Api.Infrastructure.Services.Settings.Integrations;
 
 public class DiscordSettingsProvider : IDiscordSettingsProvider
 {
